refactor(finance-ops): URL-per-tab routing
Convert Finance Ops from a single-URL client-state Tabs page to
URL-per-tab routing, matching how Facility Ops and Safety already work.

- Add financeNav.ts as the single source of tab metadata.
- Add finance-ops/layout.tsx rendering the persistent top tabs
  (router.push on change, active tab derived from usePathname) and
  enforcing the sysadmin/boardMember role gate.
- finance-ops/page.tsx now redirects to the default tab.
- Extract the inline "Pending Participants" panel into
  finance-ops/pending/page.tsx.

Behavior unchanged: same role gating, same payment-plans fetch/approve
flow, same nav badge. Bare /finance-ops links are covered by the redirect.
